Cant modify questions/groups when questionnaire is public

+++ app/controller/examiner/ajax/DeleteQuestionGroupController.php
@@ -1,30 +1,37 @@
 <?php
+	include_once '../app/model/mappers/questionnaire/QuestionnaireMapper.php';
 	include_once '../app/model/mappers/questionnaire/QuestionGroupMapper.php';
 	include_once '../app/model/mappers/questionnaire/QuestionMapper.php';
 	include_once '../app/model/mappers/questionnaire/AnswerMapper.php';
 	include_once '../app/model/mappers/actions/ParticipationMapper.php';
 
-	class DeleteQuestionGroupController extends Controller{
+	class DeleteQuestionGroupController extends Controller
+	{
 		
-		public function init(){
+		public function init()
+		{
 			$this->setOutputType( OutputType::ResponseStatus );
 			
 		}
 
-		public function run(){
+		public function run()
+		{
 			/*
 				ResponseCode
 				 0 All ok
 				 1 Authentication failed
 				 2 Access error
 				 3 Database error
+				 4 Questionnaire is public , you cant delete this.
 				-1 No Data
 			 */
-			if( isset( $_POST["question-group-id"] ) ){
+			if( isset( $_POST["question-group-id"] ) )
+			{
 
 				$participationMapper = new ParticipationMapper;
 
-				if( !($participationMapper->participatesInGroup( $_SESSION["USER_ID"] , $_POST["question-group-id"] , 2 ) || $_SESSION["USER_LEVEL"]==3)){
+				if( !($participationMapper->participatesInGroup( $_SESSION["USER_ID"] , $_POST["question-group-id"] , 2 ) || $_SESSION["USER_LEVEL"]==3))
+				{
 
 					$this->setOutput("response-code" , 2);
 					return;
@@ -33,8 +40,16 @@
 				$questionMapper = new QuestionMapper;
 				$answerMapper = new AnswerMapper;
 				$questionGroupMapper = new QuestionGroupMapper;
+				$questionnaireMapper = new QuestionnaireMapper;
 
-				try{
+				if ( $questionnaireMapper->isGroupPublic($_POST["question-group-id"]) && $_SESSION["USER_LEVEL"]!=3 )
+				{
+					$this->setOutput("response-code" , 4);
+					return;
+				}
+
+				try
+				{
 					DatabaseConnection::getInstance()->startTransaction();
 
 					$answerMapper->deleteByGroup( $_POST["question-group-id"] );
@@ -44,7 +59,8 @@
 					DatabaseConnection::getInstance()->commit();
 					$this->setOutput("response-code" , 0); 
 
-				}catch(DatabaseException $ex){
+				}catch(DatabaseException $ex)
+				{
 
 					DatabaseConnection::getInstance()->rollback();
 					$this->setOutput("response-code" , 3);
